backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── Settings & validation ─────────────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.7.0

# ── HTTP client (Ollama) ──────────────────────────────────────────────
httpx==0.28.1

# ── Image handling ────────────────────────────────────────────────────
pillow==11.0.0

# ── Word document extraction ──────────────────────────────────────────
python-docx==1.1.2

# ── Database (async SQLAlchemy + Postgres) ────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# ── Logging ───────────────────────────────────────────────────────────
python-json-logger==3.2.1

# ── Testing ───────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
